+++ lib/src/bloc/future_stated_builder.dart
@@ -1,10 +1,8 @@
 import 'package:flutter/widgets.dart';
-import 'package:stated/src/bloc/bloc.dart';
-import 'package:stated/src/core/extension/listenable.dart';
-import 'package:stated/src/core/extension/void_callback.dart';
+import 'package:stated/src/bloc/stated.dart';
 
-class FutureBlocBuilder<T> extends StatefulWidget {
-  const FutureBlocBuilder({
+class FutureStatedBuilder<T> extends StatefulWidget {
+  const FutureStatedBuilder({
     required this.future,
     required this.builder,
     this.child,
@@ -12,37 +10,44 @@ class FutureBlocBuilder<T> extends StatefulWidget {
   }) : super(key: key);
 
   final Widget? child;
-  final Future<Bloc<T>> Function(BuildContext context) future;
+  final Future<Stated<T>> Function(BuildContext context) future;
   final Widget Function(BuildContext context, T state, Widget? child) builder;
 
   @override
-  State<FutureBlocBuilder<T>> createState() => _FutureBlocBuilderState<T>();
+  State<FutureStatedBuilder<T>> createState() => _FutureStatedBuilderState<T>();
 }
 
-class _FutureBlocBuilderState<T> extends State<FutureBlocBuilder<T>> {
+class _FutureStatedBuilderState<T> extends State<FutureStatedBuilder<T>> {
   @override
   void initState() {
     super.initState();
     _initBloc();
   }
 
-  Bloc<T>? _bloc;
+  Stated<T>? _bloc;
 
   Future<void> _initBloc() async {
     final bloc = await widget.future(context);
-    bloc.subscribe(() => setState(() {})).disposeBy(bloc);
+    bloc.addListener(_notify);
     _bloc = bloc;
+
     // notify bloc is ready
-    setState(() {});
+    _notify();
   }
 
+  void _notify() => setState(() {});
+
   void _disposeBloc() async {
-    _bloc?.dispose();
-    _bloc = null;
+    if (_bloc != null) {
+      _bloc!
+        ..removeListener(_notify)
+        ..dispose();
+      _bloc = null;
+    }
   }
 
   @override
-  void didUpdateWidget(FutureBlocBuilder<T> oldWidget) {
+  void didUpdateWidget(FutureStatedBuilder<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
 
     if (oldWidget.future != widget.future ||