+++ lib/src/bloc/stated.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:stated/src/core/core.dart';
 
-abstract class Bloc<T> extends Observable implements ValueListenable<T> {
+abstract class Stated<T> extends Observable implements ValueListenable<T> {
   @protected
   @nonVirtual
   void setState([VoidCallback? callback]) {